Set unix_opendir_job C binding in its own C file
