+++ src/unix/jbuild
@@ -184,6 +184,7 @@
     unix_fstat_64_job
     unix_utimes_job
     unix_isatty_job
+    unix_opendir_job
     not_available
   ))
   (install_c_headers (
