+++ src/unix/windows_c/not_available.c
@@ -67,4 +67,5 @@ LWT_NOT_AVAILABLE1(unix_fstat_job)
 LWT_NOT_AVAILABLE1(unix_fstat_64_job)
 LWT_NOT_AVAILABLE3(unix_utimes_job)
 LWT_NOT_AVAILABLE1(unix_isatty_job)
+LWT_NOT_AVAILABLE1(unix_opendir_job)
 #endif
